prediction-engine/requirements.txt
flask
numpy
boto3
tensorflow==1.15
mysql-connector-python==8.0.5